Totally rework the rake helper to do rake loading in a fork. This is important because you can't unload code and if you load it in the current proc it will explode when trying to run it a second time. Includes a regression test for this.
The heart of the issue is that Ruby can't unload modules, so require is rebounded via $LOADED_FEATURES. You can reset that array, but then you get lots of const redefinition warnings which is gross and module level code runs twice in the same memory space which could break things. This is still way faster than actually re-exec'ing ruby and loading rake so I'll take it as a win and move on.